backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
pydantic==2.4.2
flask-migrate
